requirements.txt
Flask==3.0.0
pyserial==3.5
pygame==2.5.2
colorama==0.4.6
